requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile -U
#
-e git+https://github.com/juyrjola/django-orghierarchy.git#egg=django-orghierarchy
-e git+https://github.com/juyrjola/django-rest-framework-json-api.git#egg=djangorestframework-jsonapi
aniso8601==6.0.0
atomicwrites==1.3.0       # via pytest
attrs==19.1.0             # via packaging, pytest
backcall==0.1.0           # via ipython
beautifulsoup4==4.6.0     # via wagtail
certifi==2019.6.16        # via requests, sentry-sdk
chardet==3.0.4            # via requests
coverage==4.5.4           # via pytest-cov
decorator==4.4.0          # via ipython, traitlets
defusedxml==0.6.0         # via python3-openid, social-auth-core
django-appconf==1.0.3     # via django-image-cropping
django-cors-headers==3.0.2
django-environ==0.4.5
django-extensions==2.2.1
django-filter==2.2.0
django-helusers==0.4.6
django-image-cropping==1.2.0
django-js-asset==1.2.2    # via django-mptt
django-modelcluster==4.4  # via wagtail
django-mptt==0.10.0
django-ordered-model==3.3.0
django-summernote==0.8.11.4
django-taggit==0.24.0     # via wagtail
django-treebeard==4.3     # via wagtail
django==2.2.4
djangorestframework==3.10.2
draftjs-exporter==2.1.6   # via wagtail
drf-oidc-auth==0.9        # via django-helusers
easy-thumbnails==2.6
ecdsa==0.13.2             # via python-jose
future==0.17.1            # via pyjwkest, python-jose
graphene-django-optimizer==0.6.0
graphene-django==2.4.0
graphene==2.1.7           # via graphene-django
graphql-core==2.2.1       # via graphene, graphene-django, graphql-relay
graphql-relay==2.0.0      # via graphene
html5lib==1.0.1           # via wagtail
idna==2.8                 # via requests
importlib-metadata==0.19  # via pluggy, pytest
inflection==0.3.1
ipython-genutils==0.2.0   # via traitlets
ipython==7.7.0
jedi==0.14.1              # via ipython
lxml==4.4.0
markdown==3.1.1
more-itertools==7.2.0     # via pytest
oauthlib==3.1.0           # via requests-oauthlib, social-auth-core
packaging==19.1           # via pytest
parso==0.5.1              # via jedi
pexpect==4.7.0            # via ipython
pickleshare==0.7.5        # via ipython
pillow==6.1.0
plotly==4.1.0
pluggy==0.12.0            # via pytest
promise==2.2.1            # via graphene-django, graphql-core, graphql-relay
prompt-toolkit==2.0.9     # via ipython
psycopg2-binary==2.8.3
ptyprocess==0.6.0         # via pexpect
py==1.8.0                 # via pytest
pyasn1==0.4.6             # via rsa
pycryptodomex==3.8.2      # via pyjwkest
pygments==2.4.2           # via ipython
pyjwkest==1.4.2           # via drf-oidc-auth
pyjwt==1.7.1              # via social-auth-core
pyparsing==2.4.2          # via packaging
pytest-cov==2.7.1
pytest-django==3.5.1
pytest==5.0.1             # via pytest-cov, pytest-django
python-jose==3.0.1
python3-openid==3.1.0     # via social-auth-core
pytz==2019.2              # via django, django-modelcluster, wagtail
requests-oauthlib==1.2.0  # via social-auth-core
requests==2.22.0          # via django-helusers, pyjwkest, requests-oauthlib, social-auth-core, wagtail
retrying==1.3.3           # via plotly
rsa==4.0                  # via python-jose
rx==1.6.1                 # via graphql-core
sentry-sdk==0.10.2
singledispatch==3.4.0.3   # via graphene-django
six==1.12.0               # via django-appconf, django-extensions, django-image-cropping, graphene, graphene-django, graphql-core, graphql-relay, html5lib, packaging, plotly, promise, prompt-toolkit, pyjwkest, python-jose, retrying, singledispatch, social-auth-app-django, social-auth-core, traitlets, wagtail
social-auth-app-django==3.1.0
social-auth-core==3.2.0   # via social-auth-app-django
sqlparse==0.3.0           # via django
swapper==1.1.1
traitlets==4.3.2          # via ipython
unidecode==1.1.1          # via wagtail
urllib3==1.25.3           # via requests, sentry-sdk
wagtail==2.6.1
wcwidth==0.1.7            # via prompt-toolkit, pytest
webencodings==0.5.1       # via html5lib
willow==1.1               # via wagtail
zipp==0.5.2               # via importlib-metadata
